webapp/requirements.txt
requests==2.27.1
fastapi==0.72.0
uvicorn==0.17.0
python-dotenv==0.19.2
aiofiles==0.8.0
python-multipart==0.0.5
jinja2==3.0.3
Markdown==3.3.6
pytest==6.2.5
python-jose==3.3.0
PyYAML>=5.3.1
gunicorn==20.1.0